Refactor: SymbolList now in seperate file
