+++ src/adapters/adapter_CFL_adv.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "parser.h"
+#include "adapter.h"
 #include <GraphBLAS.h>
 #include <LAGraph.h>
