+++ src/parser.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "symbol_list.h"
 #include <LAGraph.h>
 #include <LAGraphX.h>
 
@@ -33,17 +34,6 @@ typedef struct {
     size_t rules_count;
     LAGraph_rule_EWCNF *rules;
 } grammar_t;
-
-typedef struct {
-    char *label;
-    bool is_indexed;
-    bool is_nonterm;
-} Symbol;
-
-typedef struct {
-    Symbol *symbols;
-    size_t count;
-} SymbolList;
 
 typedef struct {
     int first;
@@ -85,29 +75,7 @@ typedef struct {
     size_t valid_result;
 } config_row;
 
-typedef struct {
-    GrB_Info (*setup)(void);
-    GrB_Info (*teardown)(void);
-    GrB_Info (*init_outputs)(void);
-    GrB_Info (*free_outputs)(void);
-    GrB_Info (*run)(void);
-    GrB_Info (*prepare)(ParserResult, void *);
-    GrB_Info (*cleanup)(void);
-    bool (*is_result_valid)(size_t);
-    size_t (*get_result)(void);
-} AdapterMethods;
-
 ParserResult parser(config_row config_i);
 
-// SymbolList ans Symbol methods
-Symbol symbol_create(char *str);
-void symbol_free(Symbol sym);
-SymbolList symbol_list_create();
-void symbol_list_print(SymbolList list);
-int symbol_list_add_str(SymbolList *list, char *str, bool is_nonterm);
-int symbol_list_add_str_start_nonterm(SymbolList *list, char *str);
-void symbol_list_free(SymbolList *list);
-bool is_non_terminal(const Symbol symbol, SymbolList non_terms);
 void grammar_print(Grammar grammar, SymbolList list);
-char *read_entire_file(char *path);
 void get_configs_from_file(char *path, size_t *configs_count, config_row *configs);
